Moving ipcdef.hpp to common. Issue #230
